// frontend/src/app/applications/sync/interfaces/sync-status.interface.ts
/*
 * Copyright (C) 2012-2025 Johan Legrand <johan.legrand@sync-in.com>
 * This file is part of Sync-in | The open source file sync and share solution
 * See the LICENSE file for licensing details
 */

export interface SyncStatus {
  syncPathId: number
  state: boolean
  reportOnly: boolean
  mainError?: string
  lastErrors?: any[]
}